requirements.txt
# Core Framework
flask>=2.3.0
pydantic-ai>=0.0.14
pydantic>=2.6.0
python-dotenv>=1.0.0

# LLM Provider APIs
openai>=1.51.0
google-generativeai>=0.8.0
anthropic>=0.34.0
httpx>=0.27.0

# Flask Extensions
flask-cors>=4.0.0
werkzeug>=3.0.0
flask-session>=0.8.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pydantic-settings>=2.2.0

# Async Support
asyncio-mqtt>=0.13.0
aiofiles>=23.2.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Development
black>=24.0.0
ruff>=0.3.0
mypy>=1.9.0

# Utilities
rich>=13.7.0
typer>=0.12.0
jinja2>=3.1.0
markdown>=3.6.0 